LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigation-back-same-document-preventDefault-expected.txt


FAIL navigation.back() same-document preventDefault assert_equals: expected 1 but got 0
